Transfer learning: style-conditioned VAE, fine_tune(), versioned base checkpoint
The pretrained-base -> user-fine-tune goal, built end to end:

Model (models.py)
- FiLM style conditioning at every U-Net block and in the heightmap head;
  style 0 = vanilla base, 8 reserved custom slots, zero-init (inert at start)
- Style-noise regularisation during base training keeps the FiLM response
  smooth around the base style so fine-tuned styles land in trained territory
- U-Net deepened to 3 levels (global horizontal context at the bottleneck)
  with residual conv blocks; ~4.6M params at default width
- Encoder logvar clamped for fp16 stability

Training (train.py)
- EMA weights (warmup decay) validated and exported; AMP autocast on CUDA;
  linear-warmup + cosine LR; validation at target beta (stable early stopping)
- Biome dropout trains a reserved UNKNOWN biome row on generic terrain so
  fine-tune data from foreign worlds still conditions on something learned
- No weight decay on embeddings/norms/biases (AdamW decay would otherwise
  silently erode the frozen BASE style row during fine-tuning)
- save_base_checkpoint(): self-contained versioned dir (weights + manifest
  with exact arch + mappings)
- fine_tune(): rebuilds the model from the manifest, allocates a style slot,
  freezes the base (style row + FiLM projections trainable by default,
  ~84k params = 1.8%; "style"/"film"/"decoder" unfreeze levels), returns a
  FineTuneResult ready for export

Export + plugin
- ONNX graph gains style_id input; all styles baked into one model;
  style_mapping.json written alongside the other mappings
- Java reads declared graph inputs (older 2-input models keep working),
  loads style_mapping.json, adds /generateterrain style <name|id>,
  falls back to the UNKNOWN biome row for unmapped biomes

Data
- BiomeEncoder always reserves UNKNOWN; from_mapping()/load() rebuild a saved
  encoder verbatim for fine-tuning; grids stored as int8 (half the RAM)

Notebooks + tests
- Driver notebook: biome dropout arg, base-checkpoint step, 3-mapping export
- New FineTune.ipynb: user-facing custom-terrain walkthrough
- New test_finetune.py: UNKNOWN biome, style shapes, end-to-end fine-tune with
  bit-identical frozen-weight guarantees, ONNX input contract (9 tests green)

+++ ml/tests/test_data.py
@@ -86,7 +86,7 @@ def test_grid_shape_and_dtype(tiny):
     grouping = BlockGrouping()
     grid = csv_to_grid(tiny.data_dir / "plains1.csv", grouping, tiny)
     assert grid.shape == (tiny.chunk_width, tiny.chunk_height, tiny.chunk_depth)
-    assert grid.dtype == np.int16
+    assert grid.dtype == np.int8
     # Full chunk => no ignore voxels remain.
     assert (grid == IGNORE_INDEX).sum() == 0
